feat: add scan overlay, refactor
